requirements.txt
streamlit>=1.28.0
crewai>=0.28.0
crewai_tools>=0.0.10
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
yt-dlp>=2023.12.30 